refactor: group event-types lib functions in single folder
