+++ packages/lib/index.ts
@@ -5,5 +5,5 @@ export * from "./isBookingLimits";
 export * from "./isDurationLimits";
 export * from "./validateIntervalLimitOrder";
 export * from "./schedules";
-export * from "./getEventTypeByViewer";
-export * from "./getEventTypesPublic";
+export * from "./event-types/getEventTypeByViewer";
+export * from "./event-types/getEventTypesPublic";
